+++ lib/aws-s3.js
@@ -9,7 +9,7 @@ const objectStoreKind = cds.env.requires?.attachments?.objectStore?.kind
 const separateObjectStore = isMultitenacyEnabled && objectStoreKind === "separate"
 
 const s3ClientsCache = {}
-module.exports = class AWSAttachmentsService extends require("./basic-remote.js") {
+module.exports = class AWSAttachmentsService extends require("./object-store.js") {
   /**
    * Initializes the AWS S3 Attachments Service
    */
