+++ lib/azure-blob-storage.js
@@ -9,7 +9,7 @@ const objectStoreKind = cds.env.requires?.attachments?.objectStore?.kind
 const separateObjectStore = isMultitenacyEnabled && objectStoreKind === "separate"
 
 const azureClientsCache = {}
-module.exports = class AzureAttachmentsService extends require("./basic-remote") {
+module.exports = class AzureAttachmentsService extends require("./object-store") {
   /**
    * Initializes the Azure Blob Storage Attachments Service
    */
