+++ lib/gcp.js
@@ -9,7 +9,7 @@ const objectStoreKind = cds.env.requires?.attachments?.objectStore?.kind
 const separateObjectStore = isMultitenacyEnabled && objectStoreKind === "separate"
 
 const googleClientsCache = {}
-module.exports = class GoogleAttachmentsService extends require("./basic-remote") {
+module.exports = class GoogleAttachmentsService extends require("./object-store") {
   /**
    * Initializes the Google Cloud Platform Attachments Service
    */
