+++ lib/object-store.js
